backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
httpx>=0.24.0
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
requests>=2.31.0
aiofiles>=23.2.0
langdetect>=1.0.9
